reorganize survey widget functions, pop survey panels before showing alerts [#161]
